refactor(checker): move apparent-type display helper to dedicated module
The previous commit added apparent_intrinsic_type_display to
error_reporter/properties.rs, which pushed that file past its arch-guard size
ratchet (limit 1897). Move the shared helper into a new
error_reporter/apparent_type_display.rs module so properties.rs stays at its
ratcheted size; both call sites (element-access TS7053 in diagnostic_methods_tail
and destructuring TS2538/TS2339) reference the new module. Behavior unchanged.
